.htaccess File for URL Rewriting in Apache Server
This .htaccess file contains the URL Rewriting rules for Apache web
server.
Definition:
If any request is made to your server, which is not a file and not a
directory, that will be a Permalink request, it will be rewritten to
"rewrite.php" file, and then "rewrite.php" will handle the request.